backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Environment & Config
python-dotenv==1.0.0

# LLM & AI
google-generativeai==0.3.2

# Graph Database (Neo4j)
neo4j==5.14.0

# Vector Database (Milvus) - TODO: Uncomment when implementing vector retrieval
# pymilvus==2.3.4

# Embeddings - TODO: Uncomment when implementing vector retrieval
# sentence-transformers==2.2.2

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
